Update README.md
Reflects the simplified approach of only using the Javascript URL constructor to verify a well-formed URL.  This is because the regex found on several sites proved flawed by rejecting valid complex URLs.  Moreover it seemed on reflection to be nonsense to check the same thing in two ways so just use the one that works correctly.  Also, checking for certain aspects of the URL being present also seemed nonsense at this point, so just using the one check.